+++ asyncpg_listen/__init__.py
@@ -27,7 +27,7 @@ __all__: tuple[str, ...] = (
     "connect_func",
 )
 
-__version__ = "0.0.6"
+__version__ = "0.0.7"
 
 version = f"{__version__}, Python {sys.version}"