+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [Implement metrics/tracing on top of OpenTelemetry](https://github.com/anna-money/asyncpg-listen/pull/199)
 * [Support python 3.12 and end support of 3.8](https://github.com/anna-money/asyncpg-listen/pull/211)
+* [End support of python 3.9 and 3.10](https://github.com/anna-money/asyncpg-listen/pull/215)
 * [Eagerly start notification processing with Python 3.12+](https://github.com/anna-money/asyncpg-listen/pull/212)